LPS-61860 move classes from portal.search to portal.search.buffer to avoid package export conflicts
